Extract Rating from Ratings component
