+++ src/components/pure/Ratings.jsx
@@ -2,6 +2,7 @@ import React      from "react";
 import styled     from "styled-components";
 import Flower     from "../Flower";
 import FlowerLeaf from "../Flower/FlowerLeaf";
+import Rating     from "./Rating";
 import i18n       from "../../i18n";
 import STYLE      from "../styling/Variables";
 
@@ -42,7 +43,8 @@ const rating_groups = (ratings=[]) => {
   return groups_sorted;
 }
 
-const t = (key) => i18n.t("ratings." + key)
+const t_r = (key) => i18n.t(key)
+const t = (key) => t_r("ratings." + key)
 
 const Ratings = ({ entry, ratings, onRate, onComment }) => {
   const groups = rating_groups(ratings);
@@ -76,7 +78,7 @@ const Ratings = ({ entry, ratings, onRate, onComment }) => {
           </svg>
         </LeafWrapper>
         <RatingListForContext>
-          { g.map(r => <li key={r.id}>{Rating(r, t)}</li>) }
+          { g.map(r => <li key={r.id}>{Rating(r, t_r)}</li>) }
 
           <AdditionalCommentButtonWrapper>
             <AdditionalCommentButton onClick={() => { onComment({
@@ -127,29 +129,6 @@ const Ratings = ({ entry, ratings, onRate, onComment }) => {
   }
 }
 
-const Comment = (comment) =>
-  <div className="comment">
-    { comment.text }
-  </div>
-
-const RatingWrapper = styled.div`
-  font-size: 0.9em;
-  overflow: hidden;
-`
-
-const RatingTitle = styled.span`
-  margin-left: 0.3em;
-  font-weight: bold;
-`
-
-const RatingTitleWrapper = styled.div`
-  max-width: 288px;
-`
-
-const SourceWrapper = styled.div`
-  color: #AAA;
-  text-align: right;
-`
 
 const AdditionalCommentButtonWrapper = styled.div`
   text-align: right;
@@ -161,41 +140,6 @@ const AdditionalCommentButtonWrapper = styled.div`
 const AdditionalCommentButton = styled.button`
   float: none;
 `
-
-const rating_value_key = (value) => {
-  switch(value){
-    case -1: return "minusOne";
-    case 0: return "zero";
-    case 1: return "one";
-    case 2: return "two";
-    default: return "invalidRatingValue";
-  }
-}
-
-const Rating = (rating, t) => {
-
-  const source = rating.source ? rating.source : ''
-  const match = source.match(/(?:(?:https?|ftp):\/\/)?[a-z0-9\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF/\-?=%.]+\.[\w/\-?=%#&_.]+[\w?&%/#]+/i)
-
-  const before = match ? source.substring(0,match.index) : ''
-  const after = match ? source.substring( match.index+match[0].length, source.length)  : ''
-  let href = match ? match[0]  : ''
-  if(href.indexOf('http') !== 0)  href = 'http://' + href
-
-  let sourceSpan = () => { return match ? <span> {before} <Link target="_blank" href={href}>{t('sourceWebsite')}</Link> {after}</span> : <span>{source}</span>}
-
-  return (
-    <RatingWrapper>
-      <RatingTitleWrapper>
-        <span>{t("valueName." + rating_value_key(rating.value))}:</span><RatingTitle>{rating.title}</RatingTitle>
-      </RatingTitleWrapper>
-      <RatingCommentList>
-        {(rating.comments || []).filter(c => typeof c !== "undefined" && c !== null).map(c => <li key={c.id}>{Comment(c)}</li>)}
-      </RatingCommentList>
-      <SourceWrapper>{ sourceSpan() }</SourceWrapper>
-    </RatingWrapper>
-  )
-}
 
 module.exports = Ratings;
 
@@ -257,14 +201,6 @@ const RatingContextHeading = styled.h5`
   border-bottom: 1px solid #ddd;
   margin-bottom: 0.5em;
   width: 83%;
-`
-
-const RatingCommentList = styled.ul`
-  margin-left: 1.2em;
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
-  list-style: none;
-  padding-left: 0;
 `
 
 const RatingCount = styled.span`
